fetching Device id from the page data
